Fix epoch counter

+++ assignment1/cs231n/classifiers/neural_net.py
@@ -179,7 +179,7 @@ class TwoLayerNet(object):
     - verbose: boolean; if true print progress during optimization.
     """
     num_train = X.shape[0]
-    iterations_per_epoch = max(num_train / batch_size, 1)
+    iterations_per_epoch = int(max(num_train / batch_size, 1))
 
     # Use SGD to optimize the parameters in self.model
     loss_history = []
